source/requirements.txt
streamlit==1.32.0
pandas==2.2.1
pyyaml==6.0.1
playwright==1.42.0